gdbus: abort pending idle actions during g_dbus_connection_signal_unsubscribe()
Since commit b285899a6fb ('gdbusconnection: Document main context iteration
for unsubscriptions') the behavior during unsubscribe is better
documented.

That is, after unsubscribe we are guaranteed that the callback will no
longer be invoked (at least, if you unsubscribe while the GMainContext is
not iterated on another thread, otherwise there is obviously a race). But
it also documents, that unsubscribe does not immediately release all resources
and that only after the GDestroyNotify is called (inside the GMainContext),
that the user can be sure all resources are released. That is important
because the user must not stop iterating the GMainContext earlier or
otherwise the context and the idle actions will leak.

Note that previously, you couldn't even be sure that when scheduling an
idle action after unsubscribe, that this idle action will be invoked
before the GDestroyNotify. Instead, the GDestroyNotify gets only
scheduled after all other idle actions are processed (and they all
called signal_subscriber_unref()). If we cannot cleanup right away, it
would still be a nicer API to say that any idle action (with
G_PRIORITY_DEFAULT) scheduled after unsubscribe, will run after the
GDestroyNotify.

It would be nicer API if we could release all resources right away
during unsubscribe (at least, if we unsubscribe from inside the
GMainContext). We easily can.

Now, when calling g_dbus_connection_signal_unsubscribe() form inside the
GMainContext, then:

 - if you provide no GDestroyNotify, then that's it and all pending
   idle actions will be destroyed and released right away.

 - if you provide a GDestroyNotify, we will schedule it on an idle
   action right away.
   This means, if the user schedules another idle action after
   unsubscribe, they are guaranteed that the GDestroyNotify runs
   first.

This makes the API more strict and better behaving. As such, it should
not break any assumptions from the previous usage. Also, the
documentation is not (yet) updated to reflect the stricter behavior.